+++ CHANGELOG.md
@@ -7,9 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## Added
+### Added
 
 - Added authority claimer service to support reader mode
+- Added support to `POST` *inspect state* requests
 
 ## [1.0.0] 2023-08-22